COMP: Simplify integration of ctkIconEnginePlugin by removing proxy header
Following the update in commit b86197b7 ("COMP: Update C++ source files
removing code specific to Qt 4", 2023-07-18) and the integration of
AUTOMOC, the separate `ctkIconEnginePlugin_qt5.h` header is no longer
necessary. This change merges its content into `ctkIconEnginePlugin.h`.
